+++ src/promptflow/promptflow/_sdk/_load_functions.py
@@ -62,30 +62,40 @@ def load_common(
 
 def load_flow(
     source: Union[str, PathLike, IO[AnyStr]],
-    connections=None,
-    variant=None,
     **kwargs,
 ):
-    """Load a flow from source file. The loaded flow can be used as a function.
+    """Load flow from YAML file.
 
-    :param source: Source file path.
-    :param connections: Connections for the run.
-    :param variant: Variant of the run.
-    :param kwargs:
-    :return:
+    :param source: The local yaml source of a flow. Must be either a path to a local file.
+        If the source is a path, it will be open and read.
+        An exception is raised if the file does not exist.
+    :type source: Union[PathLike, str]
+    :return: A Flow object
+    :rtype: Flow
     """
-    flow = ProtectedFlow.load(source, **kwargs)
-    flow.context.connections = connections
-    flow.context.variant = variant
-    return flow
+    # TODO: decide whether to support setting context in here
+    return ProtectedFlow.load(source, **kwargs)
 
 
 def load_run(
     source: Union[str, PathLike, IO[AnyStr]],
+    params_override: Optional[list] = None,
     **kwargs,
 ):
+    """Load run from YAML file.
+
+    :param source: The local yaml source of a run. Must be either a path to a local file.
+        If the source is a path, it will be open and read.
+        An exception is raised if the file does not exist.
+    :type source: Union[PathLike, str]
+    :param params_override: Fields to overwrite on top of the yaml file.
+        Format is [{"field1": "value1"}, {"field2": "value2"}]
+    :type params_override: List[Dict]
+    :return: A Run object
+    :rtype: Run
+    """
     data = load_yaml(source=source)
-    return Run._load(data=data, yaml_path=source, **kwargs)
+    return Run._load(data=data, yaml_path=source, params_override=params_override, **kwargs)
 
 
 def load_connection(