+++ src/promptflow/tests/sdk_cli_azure_test/e2etests/test_cli_with_azure.py
@@ -1,28 +1,34 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+
 import os
 import sys
 import uuid
+from typing import Callable
 
 import pytest
+from azure.ai.ml.entities import Data
 
 from promptflow._cli._pf_azure.entry import main
 from promptflow._sdk.entities import Run
+from promptflow.azure import PFClient
 
 from .._azure_utils import DEFAULT_TEST_TIMEOUT, PYTEST_TIMEOUT_METHOD
 
 FLOWS_DIR = "./tests/test_configs/flows"
 DATAS_DIR = "./tests/test_configs/datas"
+RUNS_DIR = "./tests/test_configs/runs"
 
 
 # TODO: move this to a shared utility module
-def run_pf_command(*args, pf, runtime, cwd=None):
+def run_pf_command(*args, pf, runtime=None, cwd=None):
     origin_argv, origin_cwd = sys.argv, os.path.abspath(os.curdir)
     try:
         sys.argv = (
             ["pfazure"]
             + list(args)
             + [
-                "--runtime",
-                runtime,
                 "--subscription",
                 pf._ml_client.subscription_id,
                 "--resource-group",
@@ -31,6 +37,8 @@ def run_pf_command(*args, pf, runtime, cwd=None):
                 pf._ml_client.workspace_name,
             ]
         )
+        if runtime:
+            sys.argv += ["--runtime", runtime]
         if cwd:
             os.chdir(cwd)
         main()
@@ -41,9 +49,15 @@ def run_pf_command(*args, pf, runtime, cwd=None):
 
 @pytest.mark.timeout(timeout=DEFAULT_TEST_TIMEOUT, method=PYTEST_TIMEOUT_METHOD)
 @pytest.mark.e2etest
+@pytest.mark.usefixtures(
+    "mock_get_azure_pf_client",
+    "mock_set_headers_with_user_aml_token",
+    "single_worker_thread_pool",
+    "vcr_recording",
+)
 class TestCliWithAzure:
-    def test_basic_flow_run_bulk_without_env(self, pf, runtime) -> None:
-        name = str(uuid.uuid4())
+    def test_basic_flow_run_bulk_without_env(self, pf: PFClient, runtime: str, randstr: Callable[[str], str]) -> None:
+        name = randstr("name")
         run_pf_command(
             "run",
             "create",
@@ -77,9 +91,11 @@ class TestCliWithAzure:
         run = pf.runs.get(run=name)
         assert isinstance(run, Run)
 
-    def test_run_with_remote_data(self, pf, runtime, remote_web_classification_data, temp_output_dir: str):
+    def test_run_with_remote_data(
+        self, pf: PFClient, runtime: str, remote_web_classification_data: Data, randstr: Callable[[str], str]
+    ) -> None:
         # run with arm id
-        name = str(uuid.uuid4())
+        name = randstr("name1")
         run_pf_command(
             "run",
             "create",
@@ -97,7 +113,7 @@ class TestCliWithAzure:
         assert isinstance(run, Run)
 
         # run with name version
-        name = str(uuid.uuid4())
+        name = randstr("name2")
         run_pf_command(
             "run",
             "create",
@@ -113,3 +129,20 @@ class TestCliWithAzure:
         )
         run = pf.runs.get(run=name)
         assert isinstance(run, Run)
+
+    def test_run_file_with_set(self, pf: PFClient, runtime: str, randstr: Callable[[str], str]) -> None:
+        name = randstr("name")
+        run_pf_command(
+            "run",
+            "create",
+            "--file",
+            f"{RUNS_DIR}/run_with_env.yaml",
+            "--set",
+            f"runtime={runtime}",
+            "--name",
+            name,
+            pf=pf,
+        )
+        run = pf.runs.get(run=name)
+        assert isinstance(run, Run)
+        assert run.properties["azureml.promptflow.runtime_name"] == runtime